// Code/Websites/DanpheEMR/wwwroot/DanpheApp/src/app/pharmacy/report/credit-inout-patient/phrm-credit-in-out-patient-data.dto.ts
export class PHRMCreditInOutPatientReportData_DTO {
    Date: String = "";
    InvoiceNo: String = "";
    PatientCode: String = "";
    PatientName: String = "";
    Address: String = "";
    CreditAmount: number = 0;
    VisitType: String = "";
    OrganizationName: String = "";
    Remark: String = "";

}